main.py
# main.py

def select_error_correction_level(level):
    error_correction_level = {
        'L':0, 'M':1, 'Q':2, 'H':3
    }
    return error_correction_level.get(level, 1)
def select_smallest_version(error_correction_level, mode):
    versions={
        'L':{'numeric':1, 'alphanumeric':1, 'byte':1, 'kanji':1},
        'M':{'numeric':1, 'alphanumeric':1, 'byte':1, 'kanji':1},
        'Q':{'numeric':1, 'alphanumeric':1, 'byte':1, 'kanji':1},
        'H':{'numeric':1, 'alphanumeric':1, 'byte':1, 'kanji':1}
    }
    return versions[error_correction_level].get(mode, 1)
def determine_encoding_mode(data):
    if any(char.isdigit() for char in data):
        return 'numeric'
    elif any(char.isalpha() for char in data):
        return 'alphanumeric'
    elif any(char.encode('utf-8') for char in data):
        return 'byte'
    elif any(ord(char) > 127 for char in data):
        return 'kanji'
    else:
        raise ValueError("Impossible de déterminer le mode pour les données fournies.")



def determine_indicator_char_count(version, mode):
    char_count_indicators = {
        1: {'numeric': '0000', 'alphanumeric': '0000', 'byte': '0000', 'kanji': '0000'},
        9: {'numeric': '0010', 'alphanumeric': '0001', 'byte': '0000', 'kanji': '0000'}
        
    }
    return char_count_indicators[version].get(mode, '0000')


def calculate_error_correction_blocks(version, error_correction_level):
    error_correction_info = {
        'L': {1: (1, 20), 2: (1, 18), },  # Ajoutez les informations pour toutes les versions nécessaires
        'M': {1: (2, 18), 2: (2, 16), },
        'Q': {1: (2, 26), 2: (2, 24), },
        'H': {1: (4, 16), 2: (4, 12), }
    }
    return error_correction_info[error_correction_level].get(version, (0, 0))
def calculate_error_correction_words(blocks, version, error_correction_level):
    # Implémentez la logique de calcul des mots de code de correction d'erreur pour chaque bloc
    # en utilisant l'algorithme de correction d'erreur de Reed-Solomon
    pass
def implement_error_correction(encoded_data, version, error_correction_level):
    # Obtenez le nombre de blocs et de mots de code de correction d'erreur
    blocks, words_per_block = calculate_error_correction_blocks(version, error_correction_level)

    # Divisez les données encodées en blocs
    data_blocks = [encoded_data[i:i + words_per_block] for i in range(0, len(encoded_data), words_per_block)]

    # Calculez et ajoutez les mots de code de correction d'erreur pour chaque bloc
    corrected_data = ""
    for block in data_blocks:
        corrected_data += calculate_error_correction_words(block, version, error_correction_level)

    return corrected_data

def encode_data(data, mode, version, error_correction_level):
    error_correction_level = select_error_correction_level(error_correction_level)
    version = select_smallest_version(error_correction_level, mode)
    indicator_mode = determine_indicator_mode(mode)
    indicator_char_count = determine_indicator_char_count(version, mode)

    encoded_data = ""
    if mode == 'numeric':
        encoded_data += encode_numeric(data, version)
    elif mode == 'alphanumeric':
        encoded_data += encode_alphanumeric(data, version)
    elif mode == 'byte':
        encoded_data += encode_byte(data, version)
    elif mode == 'kanji':
        encoded_data += encode_kanji(data, version)

    capacity = calculate_capacity(version, error_correction_level)
    encoded_data += '0000'
    while len(encoded_data) < capacity:
        encoded_data += '0'
    encoded_data = encoded_data[:capacity]
    
    corrected_data = implement_error_correction(encoded_data, version, error_correction_level)
    
    return encoded_data

# Creation de l'image du code QR
def create_qr_matrix(version):
    matrix_size = get_matrix_size(version)
    return [[0] * matrix_size for _ in range(matrix_size)]

def create_qr_matrix_size(version):
    return 21 + 4 * (version - 1)

def place_finder_pattern(matrix):
    pass

def place_separators(matrix):
     pass
def place_alignment_pattern(matrix):
    # Cette fonction place le motif d'alignement dans la matrice du code QR
    pass

def place_timing_pattern(matrix):
    # Cette fonction place le modèle de synchronisation dans la matrice du code QR
    pass

def place_dark_module(matrix):
    # Cette fonction place le module sombre dans la matrice du code QR
    pass

def reserve_format_and_version_info(matrix):
    # Cette fonction réserve les zones d'informations sur le format et la version dans la matrice du code QR
    pass    

def main():
    data = "HELLO CC WORLD"
    mode = determine_encoding_mode(data)
    error_correction_level = 'M'
    smallest_version = select_smallest_version(error_correction_level, mode)
    
    indicator_mode = determine_indicator_mode(mode)
    indicator_char_count = determine_indicator_char_count(smallest_version, mode)
    
    
    print(f"Mode: {indicator_mode}")
    print(f"Nombre de caractères : {indicator_char_count}")
    
    encoded_data = encode_data(mode, data, smallest_version, error_correction_level)
    
    print(f"Données encodées : {encoded_data}")

if __name__ == "__main__":
    main()
